improve dir and style
add src dir
add .clang-format
add VSCode JSON files
