+++ src/config.h
@@ -1,4 +1,4 @@
-#define RV_VERSION      "1.0.1"
+#define RV_VERSION "1.0.1"
 
 /* CTRL+X: "^X"
    ALT+X: "M-X" */
@@ -38,43 +38,43 @@
 #define RVK_MARK_MOVE   "V"
 
 /* Colors available: DEFAULT, RED, GREEN, YELLOW, BLUE, CYAN, MAGENTA, WHITE, BLACK. */
-#define RVC_CWD         GREEN
-#define RVC_STATUS      CYAN
-#define RVC_BORDER      BLUE
-#define RVC_SCROLLBAR   CYAN
-#define RVC_LINK        CYAN
-#define RVC_HIDDEN      YELLOW
-#define RVC_EXEC        GREEN
-#define RVC_REG         DEFAULT
-#define RVC_DIR         DEFAULT
-#define RVC_CHR         MAGENTA
-#define RVC_BLK         MAGENTA
-#define RVC_FIFO        BLUE
-#define RVC_SOCK        MAGENTA
-#define RVC_PROMPT      DEFAULT
-#define RVC_TABNUM      DEFAULT
-#define RVC_MARKS       YELLOW
+#define RVC_CWD       GREEN
+#define RVC_STATUS    CYAN
+#define RVC_BORDER    BLUE
+#define RVC_SCROLLBAR CYAN
+#define RVC_LINK      CYAN
+#define RVC_HIDDEN    YELLOW
+#define RVC_EXEC      GREEN
+#define RVC_REG       DEFAULT
+#define RVC_DIR       DEFAULT
+#define RVC_CHR       MAGENTA
+#define RVC_BLK       MAGENTA
+#define RVC_FIFO      BLUE
+#define RVC_SOCK      MAGENTA
+#define RVC_PROMPT    DEFAULT
+#define RVC_TABNUM    DEFAULT
+#define RVC_MARKS     YELLOW
 
 /* Special symbols used by the TUI. See <curses.h> for available constants. */
-#define RVS_SCROLLBAR   ACS_CKBOARD
-#define RVS_MARK        ACS_DIAMOND
+#define RVS_SCROLLBAR ACS_CKBOARD
+#define RVS_MARK      ACS_DIAMOND
 
 /* Prompt strings for line input. */
-#define RV_PROMPT(S)    S ": "
-#define RVP_SEARCH      RV_PROMPT("search")
-#define RVP_NEW_FILE    RV_PROMPT("new file")
-#define RVP_NEW_DIR     RV_PROMPT("new dir")
-#define RVP_RENAME      RV_PROMPT("rename")
+#define RV_PROMPT(S) S ": "
+#define RVP_SEARCH   RV_PROMPT("search")
+#define RVP_NEW_FILE RV_PROMPT("new file")
+#define RVP_NEW_DIR  RV_PROMPT("new dir")
+#define RVP_RENAME   RV_PROMPT("rename")
 
 /* Number of entries to jump on RVK_JUMP_DOWN and RVK_JUMP_UP. */
-#define RV_JUMP         10
+#define RV_JUMP 10
 
 /* Default listing view flags.
    May include SHOW_FILES, SHOW_DIRS and SHOW_HIDDEN. */
-#define RV_FLAGS        SHOW_FILES | SHOW_DIRS
+#define RV_FLAGS SHOW_FILES | SHOW_DIRS
 
 /* Optional macro to be executed when a batch operation finishes. */
-#define RV_ALERT()      beep()
+#define RV_ALERT() beep()
 
 /* Shell used to launch external  programs.
    Defining   this  macro   will  force   Rover  to   launch  external
@@ -85,4 +85,4 @@
    process each time an external  program is invoked. Leave this macro
    undefined if you prefer external  programs to be launched with just
    `$EXTERNAL_PROGRAM [arg]`. */
-#define RV_SHELL        "/bin/sh"
+#define RV_SHELL "/bin/sh"
